Fix signed vs unsigned mismatch comparisons in gumbo

+++ internal/gumbo/error.c
@@ -75,7 +75,7 @@ static int print_message(GumboStringBuffer* output, const char* format, ...) {
 
 static void print_tag_stack(const GumboParserError* error, GumboStringBuffer* output) {
   print_message(output, "  Currently open tags: ");
-  for (int i = 0; i < error->tag_stack.length; ++i) {
+  for (unsigned int i = 0; i < error->tag_stack.length; ++i) {
     if (i) {
       print_message(output, ", ");
     }
@@ -156,7 +156,7 @@ static const char* find_next_newline(
 
 GumboError* gumbo_add_error(GumboParser* parser) {
   int max_errors = parser->_options->max_errors;
-  if (max_errors >= 0 && parser->_output->errors.length >= max_errors) {
+  if (max_errors >= 0 && parser->_output->errors.length >= (unsigned int)max_errors) {
     return NULL;
   }
   GumboError* error = gumbo_malloc(sizeof(GumboError));
@@ -271,7 +271,7 @@ void gumbo_init_errors(GumboParser* parser) {
 }
 
 void gumbo_destroy_errors(GumboParser* parser) {
-  for (int i = 0; i < parser->_output->errors.length; ++i) {
+  for (unsigned int i = 0; i < parser->_output->errors.length; ++i) {
     gumbo_error_destroy(parser->_output->errors.data[i]);
   }
   gumbo_vector_destroy(&parser->_output->errors);
